backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.4
pydantic-settings==2.7.1
pyyaml==6.0.2
python-dateutil==2.9.0
python-multipart==0.0.20
bcrypt==5.0.0
PyJWT==2.10.1
cryptography>=43.0.0
httpx==0.28.1
slowapi==0.1.9
